+++ pypeit/core/wavecal/templates.py
@@ -180,7 +180,7 @@ def build_template(in_files, slits, wv_cuts, binspec, outroot, outdir=None,
         #ax.plot(wave, flux)
         plt.show()
     # Generate the table
-    write_template(nwwv, nwspec, binspec, outdir, outroot, det_cut=det_cut, overwrite=overwrite)
+    wvutils.write_template(nwwv, nwspec, binspec, outdir, outroot, det_cut=det_cut, overwrite=overwrite)
 
 def grab_wvlim(kk, wv_cuts, nslits):
     """
@@ -288,47 +288,8 @@ def pypeit_identify_record(iwv_calib, binspec, specname, gratname, dispangl, out
     # Return
     return outroot
 
-
-def write_template(nwwv, nwspec, binspec, outpath, outroot, det_cut=None, order=None, overwrite=True):
-    """
-    Write the template spectrum into a binary FITS table
-
-    Args:
-        nwwv (`numpy.ndarray`_):
-            Wavelengths for the template
-        nwspec (`numpy.ndarray`_):
-            Flux of the template
-        binspec (int):
-            Binning of the template
-        outpath (str):
-        outroot (str):
-        det_cut (bool, optional):
-            Cuts in wavelength for detector snippets
-            Used primarily for DEIMOS
-        order (`numpy.ndarray`_, optional):
-            Echelle order numbers
-        overwrite (bool, optional):
-            If True, overwrite any existing file
-    """
-    tbl = Table()
-    tbl['wave'] = nwwv
-    tbl['flux'] = nwspec
-    if order is not None:
-        tbl['order'] = order
-
-    tbl.meta['BINSPEC'] = binspec
-    # Detector snippets??
-    if det_cut is not None:
-        tbl['det'] = 0
-        for dets, wcuts in zip(det_cut['dets'], det_cut['wcuts']):
-            gdwv = (tbl['wave'] > wcuts[0]) & (tbl['wave'] < wcuts[1])
-            deti = np.sum([2**ii for ii in dets])
-            tbl['det'][gdwv] += deti
-    # Write
-    outfile = os.path.join(outpath, outroot)
-    tbl.write(outfile, overwrite=overwrite)
-    print("Wrote: {}".format(outfile))
-
+def write_template():
+    pass
 
 #####################################################################################################
 #####################################################################################################
@@ -805,7 +766,7 @@ def main(flg):
         wv_vac = airtovac(wave * units.AA)
         xidl_dict = readsav(spec_file)
         flux = xidl_dict['arc1d']
-        write_template(wv_vac.value, flux, binspec, reid_path, outroot, det_cut=None)
+        wvutils.write_template(wv_vac.value, flux, binspec, reid_path, outroot, det_cut=None)
 
     # Gemini/Flamingos2
     if flg & (2**26):
